api/requirements.txt
aiosmtplib==3.0.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Deprecated==1.2.13
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
fastapi==0.110.2
filelock==3.13.4
h11==0.14.0
httpcore==1.0.5
httpx==0.26.0
idna==3.7
iniconfig==2.0.0
line_profiler==4.1.3
mailjet-rest==1.3.4
mpmath==1.3.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
passlib==1.7.4
phonenumbers==8.12.48
pkce==1.0.3
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pycryptodome==3.10.4
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pymongo==4.7.0
pytest==8.2.0
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
requests==2.31.0
requests-file==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
supertokens-python==0.18.10
sympy==1.12
text-unidecode==1.3
tldextract==3.1.0
tomli==2.0.1
twilio==7.9.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
wrapt==1.16.0
